tody up schedule manager package to follow same pattern
