+++ serverless-schedule-manager/functions/check-schedule.js
@@ -1,4 +1,4 @@
-const { prepareStudioFunction } = require(Runtime.getFunctions()['common/helpers/prepare-function'].path);
+const { prepareStudioFunction } = require(Runtime.getFunctions()['common/helpers/function-helper'].path);
 const ScheduleUtils = require(Runtime.getFunctions()['common/helpers/schedule-utils'].path);
 
 const requiredParameters = [{ key: 'name', purpose: 'name of the schedule to check' }];
